20230823 배치프로그램

+++ src/main/java/com/soccer/user/domain/KakaoProfile.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class KakaoProfile {
 
-	public Integer id;
+	public long id;
 	public String connected_at;
 	public Properties properties;
 	public KakaoAccount kakao_account;
